Fix dashboard previews
