+++ app/src/editor/shared/components/Preview/index.jsx
@@ -1,0 +1,62 @@
+// @flow
+
+import React, { type Node } from 'react'
+import cx from 'classnames'
+import ModulePreview from '$editor/shared/components/ModulePreview'
+
+import styles from './preview.pcss'
+
+type Props = {
+    aspect: {
+        height: number,
+        width: number,
+    },
+    children?: Node,
+    className?: ?string,
+    preview: any,
+    scale: number,
+    viewBoxScale: number,
+}
+
+const Preview = ({
+    aspect,
+    children,
+    className,
+    preview,
+    scale,
+    viewBoxScale,
+    ...props
+}: Props) => (
+    <svg
+        className={cx(styles.Preview, className)}
+        preserveAspectRatio="none"
+        viewBox={`0 0 ${Math.round(aspect.width * viewBoxScale)} ${Math.round(aspect.height * viewBoxScale)}`}
+        {...props}
+        style={{
+            padding: '8px',
+        }}
+    >
+        {children}
+        {preview.modules.map(({
+            height,
+            key,
+            left: x,
+            title,
+            top: y,
+            type,
+            width,
+        }) => (
+            <ModulePreview
+                height={height * scale}
+                key={key}
+                title={title || ''}
+                type={type}
+                width={width * scale}
+                x={x * scale}
+                y={y * scale}
+            />
+        ))}
+    </svg>
+)
+
+export default Preview
