+++ app/src/editor/shared/components/Preview/preview.pcss
@@ -1,0 +1,2 @@
+.root {
+}
